// src/Model/Player.ts
export interface Player{
    name: string
    ones: number
    twos: number
    threes: number
    fours: number
    fives: number
    sixes: number
    threeOfAKind: number
    fourOfAKind: number
    fullHouse: number
    smallStraight: number
    largeStraight: number
    chance: number
    kniffel: number
    total: number
    bonus: number
    upperTotal: number
    lowerTotal: number
    extraKniffel: number

}